+++ src/lib/FileType.ts
@@ -1,4 +1,7 @@
 export enum FileType {
   Brs,
+  Xml,
+  CodeBehind,
+  ViewXml,
   Other,
 }